Create no-data-privacy-policy
Privacy Policy
===

By using our apps and services you are consenting to our policies regarding the collection, use and disclosure of personal information set out in this privacy policy.

Collection of Personal Information
---

We are not interested in collecting any personal information. We believe such information is yours and yours alone.
We do not collect, use, store, share or transmit any information, personal or otherwise, nor do we include any advertising or analytics software that talks to third parties.

### Email
However, if you email the developer for support or other feedback, the emails with email addresses may be retained in their mailbox (and email provider's systems). The email addresses will be used only to reply to the concerns or suggestions raised.

### Disclosure of Personal Information

We will not disclose your information to any third party except if you expressly consent or where required by law.

### Contacting Us

If you have any questions or concerns, please feel free to email questions@owsega.dev